Getting data from auth0
