+++ src/handlers/users/list.js
@@ -1,65 +1,30 @@
 'use strict';
 
-const fetch = require('node-fetch');
-const config = require('../../config/constants.js');
 const authMiddleware = require('../../middlewares/auth0.js');
 
-// Get an access token for the Auth0 Admin API
-function getAdminAccessToken() {
-  const options = {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify({
-      client_id: config.auth0.CLIENT_ID,
-      client_secret: config.auth0.CLIENT_SECRET,
-      audience: `${config.auth0.DOMAIN}/api/v2/`,
-      grant_type: 'client_credentials',
-    }),
-  };
-
-  return fetch(`${config.auth0.DOMAIN}/oauth/token`, options)
-    .then((response) => response.json());
-}
-
-
-// Get the user's profile from the Admin API
-function getUserProfile(accessToken, userID) {
-  const options = {
-    headers: {
-      'Authorization': `Bearer ${accessToken}`
-    }
-  }
-  return fetch(`${config.auth0.DOMAIN}/api/v2/users/${userID}`, options)
-    .then(response => response.json());
-}
-
-module.exports.list = authMiddleware(async (context, request) => {
-  let res = {};
-  try {
-    const data = await getAdminAccessToken();
-    const user = await getUserProfile(data.access_token, request.user.sub);
-
-    res = {
-      body: {
-        success: true,
-        user,
-      },
-    };
-  } catch (error) {
-    res = {
-      status: 500,
-      body: {
-        success: false,
-        error,
-      },
-    };
-  }
+module.exports.list = authMiddleware((context) => {
+  // TODO: get this data from the database
+  const users = [{
+    id: 1,
+    name: 'Emma Wedekind'
+  }, {
+    id: 2,
+    name: 'Crysfel Villa'
+  }, {
+    id: 3,
+    name: 'Mosh Feu'
+  }];
 
   context.res = {
-    ...res,
+    status: 200,
+    body: {
+      success: true,
+      users,
+    },
     headers: {
       'Content-Type': 'application/json'
-    },
+    }
   };
-  context.done()
+
+  context.done();
 });
